+++ lib/src/cook/process/toolchain/serialize/MSVC.cpp
@@ -91,12 +91,16 @@ namespace cook { namespace process { namespace toolchain { namespace serialize {
         oss << "" << std::endl;
         oss << "    tm[Part.Cli]            = fun(k,v) { return k }" << std::endl;
         oss << "    tm[Part.Pre]            = fun(k,v) { if (v.empty) { return \"/${k}\" } else { return \"/${k}:${v}\" } }" << std::endl;
+        oss << "    tm[Part.Export]         = fun(k,v) { return \"/DEF:${k}\" }" << std::endl;
         oss << "    tm[Part.Output]         = fun(k,v) { return \"/OUT:${k}\" }" << std::endl;
         oss << "    tm[Part.Input]          = fun(k,v) { return k }" << std::endl;
         oss << "    tm[Part.Library]        = fun(k,v) { return \"${k}.lib\" }" << std::endl;
         oss << "    tm[Part.LibraryPath]    = fun(k,v) { if (k.empty) { return \"/LIBPATH:./\" } else { return \"/LIBPATH:${k}\" } }" << std::endl;
         oss << "    " << std::endl;
         oss << "    kv.append(Part.Cli, \"link\")" << std::endl;
+        oss << "    if (s == TargetType.SharedLibrary){" << std::endl;
+        oss << "        kv.append(Part.Pre, \"DLL\")" << std::endl;
+        oss << "    }" << std::endl;
         oss << "}" << std::endl;
     }